hackathon/requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-mail-templated==2.6.5
django-sendgrid-v5==0.8.0
future==0.17.1
gunicorn==19.9.0
Pillow==6.1.0
python-http-client==3.1.0
pytz==2019.1
sendgrid==6.0.5
sqlparse==0.3.0
whitenoise==4.1.3
